requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coverage==7.3.0
cryptography==41.0.3
Django
django-auto-logout
django-autoslug
django-autoslug-field
django-countries
django-debug-toolbar
django-dirtyfields
django-extensions
django-js-asset
django-mptt
django-nose
django-phone-field
django-phonenumber-field
django-plugins
django-tree
django-treebeard
Faker==20.1.0
flake8==6.1.0
git-filter-repo==2.38.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
nose==1.3.7
packaging==23.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
phonenumbers==8.13.23
Pillow==10.0.0
pluggy==1.3.0
psycopg2==2.9.7
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
pyOpenSSL==23.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==6.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
